Refactor code (clean architecture)
